requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
redis==5.1.0
pydantic==2.9.0
pydantic-settings==2.5.0
anthropic==0.40.0
elasticsearch[async]==8.15.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
structlog==24.4.0
pyyaml==6.0.2
sentry-sdk[fastapi]==2.14.0
networkx==3.3
grpcio==1.78.0
grpcio-tools==1.78.0
grpcio-reflection==1.78.0
pgvector==0.3.6
sentence-transformers==3.4.1
numpy==2.2.6
pandas==2.2.3
